Put initial state in seperate file
